refactor(features): change feature split structure (#107)
