+++ src/components/templates/MainLayout.tsx
@@ -5,10 +5,12 @@ import CookieBar from '@/components/organisms/CookieConsent';
 import Footer from '@/components/organisms/Footer';
 import Header from '@/components/organisms/Header';
 import { ChatBubbleHost } from '@/features/chat_bubble';
-import { DeadPixelHost } from '@/features/dead_pixel';
 import { ContainerGiftFlaps } from '@/features/gifts';
 import { AdblockerSuspectBar } from '@/features/gifts';
-import { StickyVideoExperienceHost } from '@/features/sticky_video';
+import {
+  DeadPixelHost,
+  StickyVideoExperienceHost,
+} from '@/features/obstruction_decor';
 import { WheelOfFortuneHost } from '@/features/wheel_of_fortune';
 import { useExperienceFlagsStore } from '@/lib/state/experience_flags';
