+++ src/features/browser_core/index.ts
@@ -1,2 +1,3 @@
+export { default as useContextMenu } from './hooks/useContextMenu';
 export { default as PageTitleExperienceHost } from './PageTitleExperienceHost';
 export { default as string_marquee } from './utils/string_marquee';
