+++ src/features/browser_core/utils/string_marquee.ts
