+++ src/features/content/components/CoverPlaceholder.tsx
