+++ src/features/content/components/PartitionalLockedContent.tsx
