+++ src/features/content/components/SmallCoverListItem.tsx
@@ -5,7 +5,7 @@ import { FunctionComponent } from 'react';
 
 import { CoverPlaceholder } from './CoverPlaceholder';
 
-import { type ArticleDatum } from '@/features/articles';
+import { type ArticleDatum } from '@/features/content';
 
 type SmallCoverListItemProps = {
   article: ArticleDatum;
