+++ src/features/content/index.ts
@@ -5,3 +5,4 @@ export { LargeCoverItem } from './components/LargeCoverItem';
 export { SmallCoverListItem } from './components/SmallCoverListItem';
 export { TextListItem } from './components/TextListItem';
 export { ArticleServiceSingleton as ArticleService } from './services/ArticleService';
+export { default as PartitionalLockedContent } from './components/PartitionalLockedContent';
