--- src/features/content_limiter/index.ts
@@ -1,1 +1,0 @@
-export { default as PartitionalLockedContent } from './components/PartitionalLockedContent';
