--- src/features/context_menu/index.ts
@@ -1,1 +1,0 @@
-export { default as useContextMenu } from './hooks/useContextMenu';
