--- src/features/dead_pixel/index.ts
@@ -1,1 +1,0 @@
-export { default as DeadPixelHost } from './DeadPixelHost';
