+++ src/features/obstruction_decor/DeadPixelHost.tsx
