+++ src/features/obstruction_decor/StickyVideoExperienceHost.tsx
