+++ src/features/obstruction_decor/index.ts
@@ -1,1 +1,2 @@
 export { default as StickyVideoExperienceHost } from './StickyVideoExperienceHost';
+export { default as DeadPixelHost } from './DeadPixelHost';
